requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs
google-cloud-bigquery<3.0.0
fastapi
uvicorn

# utilities
six>=1.14
joblib
memoized-property
termcolor
python-dotenv
direnv

#beautiful soup
beautifulsoup4

#data-transformers not useful for package just for one notebook
#datasets

#tests
httpx

#preprocessing
nltk
contractions

# deep learning
tensorflow==2.8
transformers
